smdk4412-common: camera: implement resizing preview aspect ratio
S5C73M3 Back camera provides a preview frame and a JPEG image in
the form of an interleaved buffer. However, the downside is that
the sensor returns invalid buffer whenever the preview aspect ratio
doesnt match the aspect ratio of the JPEG picture. This breaks camera
capture on applications which use different aspect ratio for preview and
capture.

There is currently no proper way to enforce these aspect ratio restrictions
since the camera implementations differ between applications. As a
workaround, we switch the preview size in the HAL based on the Picture
size so that the aspect ratios always match. This may cause stretching of
the preview since the Surface on the Camera app maybe configured for a
different size, but a functioning image capture is more important.

Change-Id: Ieedcd407b0525f9925121af1cf7d71076cfb10d9